refactor(core): move more responsibility to workspace routing
